Remove multi-line comments from python code
This commit adds support for multi-line comments in the script. Previously, the script only handled single-line comments starting with '#' and "//". With this change, the script can now also handle multi-line python comments enclosed within triple quotes ('"""'). This is achieved by introducing a new boolean variable 'inComment' to keep track of whether the current line is inside a multi-line comment or not. The script then skips lines that are inside a multi-line comment.